+++ domComponent.js
@@ -1,37 +1,37 @@
 var createElement = require('virtual-dom/create-element');
 var diff = require('virtual-dom/diff');
 var patch = require('virtual-dom/patch');
-var coerceChildren = require('./coerceChildren');
-var isVnode = require('virtual-dom/vnode/is-vnode');
-var isWidget = require('virtual-dom/vnode/is-widget');
+var toVdom = require('./toVdom');
+var isVdom = require('./isVdom');
 
 function DomComponent(options) {
   this.document = options && options.document;
 }
 
-function checkVdom(vdom) {
-  if (!isVnode(vdom) && !isWidget(vdom)) {
+function prepareVdom(object) {
+  var vdom = toVdom(object);
+  if (!isVdom(vdom)) {
     throw new Error('expected render to return vdom');
+  } else {
+    return vdom;
   }
 }
 
 DomComponent.prototype.create = function (vdom) {
-  this.vdom = coerceChildren.toVdom(vdom);
-  checkVdom(this.vdom);
+  this.vdom = prepareVdom(vdom);
   return this.element = createElement(this.vdom, {document: this.document});
 };
 
 DomComponent.prototype.merge = function (vdom, element) {
-  this.vdom = coerceChildren.toVdom(vdom);
-  checkVdom(this.vdom);
+  this.vdom = prepareVdom(vdom);
   return this.element = element;
 };
 
 DomComponent.prototype.update = function (vdom) {
-  var patches = diff(this.vdom, coerceChildren.toVdom(vdom));
-  this.element = patch(this.element, patches);
-  this.vdom = vdom;
-  return this.element;
+  var oldVdom = this.vdom;
+  this.vdom = prepareVdom(vdom);
+  var patches = diff(oldVdom, this.vdom);
+  return this.element = patch(this.element, patches);
 };
 
 DomComponent.prototype.destroy = function (options) {
